Pushed optimise method in subclass up into SectorModel
